@@ -1,38 +1,38 @@
-const sensor = require('./sensor');
+const sensor = require('./stimulus');
 
-var StimulusListener = function() {
-  this._sensor = sensor;
-  this._stimulusId = sensor.getNextId();
-}
-
-StimulusListener.prototype.listenTo = function(triggerName, defaultValue) {
-  if (defaultValue !== null && typeof defaultValue !== 'undefined') {
-    this[triggerName] = defaultValue;
+export default class Listener {
+  constructor () {
+    this._sensor = sensor;
+    this._stimulusId = sensor.getNextId();
   }
-  this._sensor.listenTo(this, triggerName);
-}
 
-StimulusListener.prototype.follow = function(triggerName, defaultValue) {
-  if (defaultValue !== null && typeof defaultValue !== 'undefined') {
-    this[triggerName] = defaultValue;
+  listenTo (triggerName, defaultValue) {
+    if (defaultValue !== null && typeof defaultValue !== 'undefined') {
+      this[triggerName] = defaultValue;
+    }
+    this._sensor.listenTo(this, triggerName);
   }
-  this._sensor.follow(this, triggerName);
-}
 
-StimulusListener.prototype.fire = function(triggerName, args) {
-  this._sensor.fire(triggerName, args);
-}
+  follow (triggerName, defaultValue) {
+    if (defaultValue !== null && typeof defaultValue !== 'undefined') {
+      this[triggerName] = defaultValue;
+    }
+    this._sensor.follow(this, triggerName);
+  }
 
-StimulusListener.prototype.propagate = function(triggerName, data, args) {
-  this._sensor.propagate(triggerName, data, args);
-}
+  fire (triggerName, args) {
+    this._sensor.fire(triggerName, args);
+  }
 
-StimulusListener.prototype.dataReceiver = function(trigger) {
-  this[trigger.name] = trigger.data;
-}
+  propagate (triggerName, data, args) {
+    this._sensor.propagate(triggerName, data, args);
+  }
 
-StimulusListener.prototype.shouldReceiveData = function(trigger) {
-  return true;
-}
+  dataReceiver (trigger) {
+    this[trigger.name] = trigger.data;
+  }
 
-module.exports = StimulusListener;
+  shouldReceiveData (trigger) {
+    return true;
+  }
+}
